requirements.txt
streamlit>=1.28.0
groq>=0.4.0
langgraph>=0.0.60
langchain>=0.1.0
langchain-groq>=0.1.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
